+++ packages/moonstone/Scroller/Scrollbar.js
@@ -180,7 +180,6 @@ class ScrollbarBase extends PureComponent {
 	minThumbSizeRatio = 0
 	pressed = false
 	ignoreMode = false
-	buttonToFocus = null
 	// component refs
 	containerRef = null
 	thumbRef = null
@@ -189,10 +188,6 @@ class ScrollbarBase extends PureComponent {
 
 	setPressStatus = (isPressed) => {
 		this.pressed = isPressed;
-		if (!isPressed && this.buttonToFocus) {
-			Spotlight.focus(this.buttonToFocus);
-			this.buttonToFocus = null;
-		}
 	}
 
 	setIgnoreMode = (shouldIgnore) => {
@@ -217,7 +212,7 @@ class ScrollbarBase extends PureComponent {
 			maxPos = vertical ? bounds.maxTop : bounds.maxLeft,
 			shouldDisablePrevButton = currentPos <= 0,
 			shouldDisableNextButton = currentPos >= maxPos,
-			spotItem = window.document.activeElement;
+			spotItem = Spotlight.getCurrent();
 
 		this.setState((prevState) => {
 			const
@@ -233,16 +228,11 @@ class ScrollbarBase extends PureComponent {
 			}
 		});
 
-		if (shouldDisablePrevButton && spotItem === prevButtonNodeRef) {
-			if (this.pressed) {
-				this.setIgnoreMode(true);
-				this.buttonToFocus = nextButtonNodeRef;
-			}
-		} else if (shouldDisableNextButton && spotItem === nextButtonNodeRef) {
-			if (this.pressed) {
-				this.setIgnoreMode(true);
-				this.buttonToFocus = prevButtonNodeRef;
-			}
+		if (this.pressed && (
+			shouldDisablePrevButton && spotItem === prevButtonNodeRef ||
+			shouldDisableNextButton && spotItem === nextButtonNodeRef
+		)) {
+			this.setIgnoreMode(true);
 		}
 	}