+++ packages/moonstone/CHANGELOG.md
@@ -24,6 +24,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/FormCheckbox` to display correct colors in light skin
 - `moonstone/Picker` and `moonstone/RangePicker` to forward `onKeyDown` events when not `joined`
 - `moonstone/SelectableItem` to display correct icon width and alignment
+- `moonstone/Scroller` to properly 5-way navigate from scroll buttons
 
 ## [1.6.1] - 2017-08-07